chore(confd): bump confd to v0.4.1
There are a few fixes in confd that will better help us resolve issues like deis/deis#499. Bumping
to v0.4.1 because v0.4.0 had a small bug where the --backend option was mandatory (see
kelseyhightower/confd#86). This bump should be completely backwards compatible and no changes
should be required in the containers to work with this version.

Make sure that when you rebuild the containers for testing that confd is running inside the
container and isn't making a fuss.

AFFECTED COMPONENTS: The following components use confd and are therefore affected by this change:

- builder
- cache (thought doesn't actively use it; this should be refactored)
- controller
- database
- registry
- router

The only component that is not affected is the logger.